pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icesculpt"
version = "0.1.0"
description = "The art of theme crafting for IceWM — a visual editor for shaping your desktop identity."
readme = "README.md"
license = "LGPL-2.1-or-later"
requires-python = ">=3.8"
authors = [
    { name = "IceSculpt Contributors" },
]
keywords = ["icewm", "theme", "editor", "gtk", "window-manager", "xpm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Desktop Environment",
    "Topic :: Multimedia :: Graphics :: Editors",
]
dependencies = [
    "PyGObject>=3.30",
    "pycairo>=1.16",
]

[project.urls]
Homepage = "https://github.com/nicholasgasior/icesculpt"
Issues = "https://github.com/nicholasgasior/icesculpt/issues"

[project.scripts]
icesculpt = "icesculpt.app:main"

[project.gui-scripts]
icesculpt = "icesculpt.app:main"

[tool.setuptools.packages.find]
include = ["icesculpt*"]

[tool.setuptools.package-data]
icesculpt = [
    "data/*.json",
    "data/presets/*.json",
    "data/pixmap_templates/*",
    "data/icons/*.svg",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
